+++ gpu/src/lib.rs
@@ -6,3 +6,6 @@ If you were using the `--all-features` flag please read this crate's Cargo.toml"
 
 #[cfg(feature = "metal")]
 pub mod metal;
+
+#[cfg(feature = "cuda")]
+pub mod cuda;
