+++ math/src/gpu/cuda/mod.rs
@@ -1,2 +1,1 @@
-pub mod abstractions;
 pub mod fft;
